+++ CHANGELOG.md
@@ -27,3 +27,17 @@ It can return random values ​​for String and Integer types, depending on the
 ### **JeaWidget**
 
 Use this if you want to add new widgets to your app with many ready made schemes!
+
+
+## 0.0.3
+
+## **New Packages!**
+
+* `JeaFriday`
+
+## **New Features!**
+
+* `JeaWidget`: JeaHorizontalScrollView, JeaVerticalScrollView.
+* `JeaRandom`: Random Email, Random Password.
+* `JeaFriday`: to (Navigator), back (Navigator), go (Navigator), allCloseAndGo (Navigator), getRequestContent, getRequestHeaders, getRequestExtra.
+* `JeaFire`: get.